requirements.txt
streamlit
boto3
bedrock-agentcore
strands-agents
python-dotenv
